requirement.txt
arabic-reshaper==2.1.0
asgiref==3.2.10
asttokens==2.0.5
beautifulsoup4==4.10.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==8.0.3
colorama==0.4.4
coreapi
coreschema==0.0.4
coverage==4.0.3
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.7
django-appconf==1.0.5
django-cors-headers==3.5.0
django-imagekit==4.0.2
django-import-export==2.4.0
django-jazzmin==2.2.9
django-safedelete==0.5.6
django-softdelete
django-taggit==1.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-generators==0.5.0
easy-password-generator==0.0.7
et-xmlfile==1.0.1
executing==0.8.2
exponent-server-sdk
Faker==4.1.3
fpdf==1.7.2
future==0.18.2
html5lib==1.1
icecream==2.1.1
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.5
pilkit==2.0
Pillow==8.4.0
pip-tools==4.4.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
PyMySQL==0.10.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
reportlab==3.6.2
sentry-sdk==1.1.0
six==1.14.0

soupsieve==2.3
sqlparse==0.3.0
tablib==2.0.0
text-unidecode==1.3
tinycss2==1.0.2
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
geographiclib==1.52
geopy==2.1.0
stripe
django-paypal

google-currency
django-user-agents==0.4.0
drf-api-tracking==1.8.0
icecream==2.1.1
WeasyPrint==52.5